renamed default settings to default_settings.py
