+++ docs/searchsetup.rst
@@ -7,23 +7,23 @@ Tardis comes with a pluggable single search option which provides users with a s
 which can take anything from a simple text search to full lucene query syntax and return a list
 of matching Experiment, datasets and datafiles.
 
-The single search box uses SOLR, with a haystack frontent, and accordingly requires some setup. 
+The single search box uses SOLR, with a haystack frontent, and accordingly requires some setup.
 The single search box is disabled by default.
 
 Setting up SOLR
 ===============
-SOLR doesn't work out of the box with MyTardis. It is not currently installed when buildout is run and requires a number of manual steps to get working. 
+SOLR doesn't work out of the box with MyTardis. It is not currently installed when buildout is run and requires a number of manual steps to get working.
 
-SOLR runs as a Java applet and can be served from any suitable container. We use Tomcat here. 
+SOLR runs as a Java applet and can be served from any suitable container. We use Tomcat here.
 
-The following are a very simple list of steps that will get everything up and running. It is advisable to follow up with the person responsible for overseeing security policy at your home institution to see if any extra setup is necessary 
+The following are a very simple list of steps that will get everything up and running. It is advisable to follow up with the person responsible for overseeing security policy at your home institution to see if any extra setup is necessary
 
 Tomcat
 ------
 
-Tomcat 6.0 (tested with the code) can be downloaded from 
+Tomcat 6.0 (tested with the code) can be downloaded from
 
-http://apache.mirror.aussiehq.net.au/tomcat/tomcat-6/v6.0.33/bin/apache-tomcat-6.0.33.tar.gz .  
+http://apache.mirror.aussiehq.net.au/tomcat/tomcat-6/v6.0.33/bin/apache-tomcat-6.0.33.tar.gz .
 
 Un-tar the package and move it to some appropriate location. We use /usr/local/tomcat6 in this example. Please talk to your sysadmin for instructions for your home institution.
 
@@ -57,7 +57,7 @@ Copy and paste the following in solr.xml
     <Context docBase="/usr/local/tomcat6/webapps/solr.war" debug="0" crossContext="true" >
        <Environment name="solr/home" type="java.lang.String" value="/usr/local/tomcat6/solr" override="true" />
     </Context>
-   
+
 Start Tomcat using
 
     sudo /usr/local/tomcat6/bin/statup.sh
@@ -74,11 +74,11 @@ Django Configuration
 Enabling Single Search
 ----------------------
 
-A list of default settings for Single Search are already in settings_changeme.py in the MyTardis repository. Single search is enabled by setting the SINGLE_SEARCH_ENABLED option to True.
+A list of default settings for Single Search are already in default_settings.py in the MyTardis repository. Single search is enabled by setting the SINGLE_SEARCH_ENABLED option to True.
 
 Other settings are shown below:
 
-HAYSTACK_SITECONF 
+HAYSTACK_SITECONF
 
 Points to the location of the search_sites.py file that Haystack looks to for instructions on which models to index. Defaults to 'tardis.search_sites'
 
@@ -97,7 +97,7 @@ Whether or not Haystack is allowed to register signals with Django. This will di
 Creating a Configuration
 ------------------------
 
-Haystack can automatically generate a SOLR schema file based on the indexes defined in the Haystack indexes file (tardis/tardis_portal/search_indexes.py). The schema file is a description of how items added to the index are processed during indexing, and how they may be searched. 
+Haystack can automatically generate a SOLR schema file based on the indexes defined in the Haystack indexes file (tardis/tardis_portal/search_indexes.py). The schema file is a description of how items added to the index are processed during indexing, and how they may be searched.
 
 A configuration is generated using the following command from the main directory of a deployed MyTardis checkout:
 
@@ -137,7 +137,7 @@ Note: Changes to the structure or properties of models and schemas (as opposed t
 Updating Schema
 ---------------
 
-SOLR depends on two types of information: Information about the fields of a model, and how they are to be added to the search index (reflected in schema.xml) and the actual data contained in these models which (indexed as documents in the search index). When new data is added or changed, this can easily be reflected in the index. However, the schema.xml file is a static representation of what SOLR expects objects for indexing to look like. If changes are made to the structure of models (e.g. new fields added, field types changed) or data schemas are added, removed or changed, then the static representation of data in schema.xml will differ from the data being passed into SOLR. 
+SOLR depends on two types of information: Information about the fields of a model, and how they are to be added to the search index (reflected in schema.xml) and the actual data contained in these models which (indexed as documents in the search index). When new data is added or changed, this can easily be reflected in the index. However, the schema.xml file is a static representation of what SOLR expects objects for indexing to look like. If changes are made to the structure of models (e.g. new fields added, field types changed) or data schemas are added, removed or changed, then the static representation of data in schema.xml will differ from the data being passed into SOLR.
 
 Given that the editing of data schemas is a core function of MyTardis, we need a way to update the schema.xml file when changes are made to MyTardis data schemas. This is not supported by default with Haystack, so we have written a small batch script which will check for changes in the schema.xml file generated by the build_solr_schema admin command. If there are changes, it will replace the existing schema.xml file, restart SOLR and then rebuild the indexes from scratch.
 
@@ -151,4 +151,4 @@ Adding to templates
 
 Single Seach adds a 'search_form' variable to all contexts once enabled and working (it actually does this always but the variable will be set to None when SingleSearch is enabled). This form can be rendered to your main portal template to provide a persistent search box across all views of your tardis deployment. The form returns results in a special search view specifically for search results.
 
-Single Search also adds some extra functionality to a number of the existing templates (specifically the experiment_view.html template and a number of its ajax templates) to add highlighting to results based on search results. Care should be taking if overriding these templates in your own deployment. 
+Single Search also adds some extra functionality to a number of the existing templates (specifically the experiment_view.html template and a number of its ajax templates) to add highlighting to results based on search results. Care should be taking if overriding these templates in your own deployment.
